API for ReactionDiffusion.jl

TOM WILL DO THIS: explain philosophy of package - easy-to-code, well-chosen defaults, fast, optimized. Built on SciML. Not as modular and extensible.

Interoperability with SciML packages

TOM WILL DO THIS: Speak about interoperability with Catalyst, MTK, DiffEQ

Interoperability with plotting packages

TOM WILL DO THIS: Plots and Makie

Data structures

Functions

Exported functions

ReactionDiffusion.screen_values — Function

screen_values(;min=0,max=1,number=10, mode="linear")

Return double the number x plus 1.

Add lots more docstrings to the scripts in src/.

Functions that are not exported but can be loaded